requirements.txt
flask==2.3.2
flask-socketio==5.3.6
opencv-python==4.8.0.76
ultralytics==8.0.196
pymongo==4.5.0
pillow==10.0.1
werkzeug==2.3.7
pywhatkit==5.4
flask-cors==4.0.0
deepface==0.0.79
tensorflow==2.13.0
numpy==1.24.4
streamlit==1.27.0
python-dotenv==1.0.0